+++ lib/dart_odbc.dart
@@ -1,5 +1,5 @@
 export 'src/dart_odbc.dart';
-export 'src/dart_odbc_base.dart';
+export 'src/dart_odbc_base/base.dart';
 export 'src/exceptions.dart';
 export 'src/helper.dart';
 export 'src/libodbc.dart';