+++ lib/src/dart_odbc_base/init.dart
@@ -1,0 +1,34 @@
+part of './base.dart';
+
+extension on DartOdbc {
+  void _initialize() {
+    final pHEnv = calloc<SQLHANDLE>();
+    tryOdbc(
+      _sql.SQLAllocHandle(SQL_HANDLE_ENV, nullptr, pHEnv),
+      operationType: SQL_HANDLE_ENV,
+      handle: pHEnv.value,
+      onException: HandleException(),
+      beforeThrow: () {
+        calloc.free(pHEnv);
+      },
+    );
+    _hEnv = pHEnv.value;
+
+    tryOdbc(
+      _sql.SQLSetEnvAttr(
+        _hEnv,
+        SQL_ATTR_ODBC_VERSION,
+        Pointer.fromAddress(SQL_OV_ODBC3),
+        0,
+      ),
+      handle: _hEnv,
+      operationType: SQL_HANDLE_ENV,
+      onException: HandleException(),
+      beforeThrow: () {
+        calloc.free(pHEnv);
+      },
+    );
+
+    calloc.free(pHEnv);
+  }
+}