+++ test/async_test.dart
@@ -12,8 +12,8 @@ void main() {
   test('parallel async query test', () async {
     final futures = List.generate(10, (index) {
       return helper.query(
-      'SELECT ? AS index_value;',
-      params: [index],
+        'SELECT ? AS index_value;',
+        params: [index],
       );
     });